Moving num_mwm_id.hpp to routing_common.
