+++ map/framework.cpp
@@ -13,12 +13,13 @@
 #include "private.h"
 
 #include "routing/index_router.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/online_absent_fetcher.hpp"
 #include "routing/road_graph_router.hpp"
 #include "routing/route.hpp"
 #include "routing/routing_algorithm.hpp"
 #include "routing/routing_helpers.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "search/downloader_search_callback.hpp"
 #include "search/editor_delegate.hpp"
