+++ map/routing_helpers.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "routing/num_mwm_id.hpp"
+#include "routing_common/num_mwm_id.hpp"
 
 #include "storage/country_info_getter.hpp"
 #include "storage/storage.hpp"
