+++ routing/cross_mwm_index_graph.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
 #include "routing/cross_mwm_connector.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/segment.hpp"
 #include "routing/routing_exceptions.hpp"
 #include "routing/transition_points.hpp"
 #include "routing/vehicle_mask.hpp"
 
+#include "routing_common/num_mwm_id.hpp"
 #include "routing_common/vehicle_model.hpp"
 
 #include "coding/file_container.hpp"
