+++ routing/index_graph_loader.hpp
@@ -2,9 +2,9 @@
 
 #include "routing/edge_estimator.hpp"
 #include "routing/index_graph.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/vehicle_mask.hpp"
 
+#include "routing_common/num_mwm_id.hpp"
 #include "routing_common/vehicle_model.hpp"
 
 #include "indexer/index.hpp"
