+++ routing/index_graph_starter.hpp
@@ -7,11 +7,12 @@
 #include "routing/fake_vertex.hpp"
 #include "routing/index_graph.hpp"
 #include "routing/joint.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/route_point.hpp"
 #include "routing/route_weight.hpp"
 #include "routing/segment.hpp"
 #include "routing/world_graph.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include <cstddef>
 #include <cstdint>
