+++ routing/index_road_graph.hpp
@@ -1,9 +1,10 @@
 #pragma once
 
 #include "routing/index_graph_starter.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/road_graph.hpp"
 #include "routing/segment.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "indexer/index.hpp"
