+++ routing/pedestrian_directions.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
 #include "routing/directions_engine.hpp"
-#include "routing/num_mwm_id.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include <memory>
