+++ routing/segment.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
-#include "routing/num_mwm_id.hpp"
 #include "routing/road_point.hpp"
 #include "routing/route_weight.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include <cstdint>
 #include <sstream>
