+++ routing/single_vehicle_world_graph.hpp
@@ -5,10 +5,11 @@
 #include "routing/geometry.hpp"
 #include "routing/index_graph.hpp"
 #include "routing/index_graph_loader.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/road_graph.hpp"
 #include "routing/segment.hpp"
 #include "routing/world_graph.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "geometry/point2d.hpp"
