+++ routing/traffic_stash.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
-#include "routing/num_mwm_id.hpp"
 #include "routing/segment.hpp"
 
 #include "traffic/traffic_cache.hpp"
 #include "traffic/traffic_info.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "indexer/mwm_set.hpp"
