+++ routing/transit_graph_loader.hpp
@@ -2,8 +2,9 @@
 
 #include "routing/edge_estimator.hpp"
 #include "routing/index_graph.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/transit_graph.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "indexer/index.hpp"
