+++ track_analyzing/serialization.hpp
@@ -2,8 +2,9 @@
 
 #include "track_analyzing/track.hpp"
 
-#include "routing/num_mwm_id.hpp"
 #include "routing/segment.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "platform/country_file.hpp"
