+++ track_analyzing/track.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "routing/num_mwm_id.hpp"
 #include "routing/segment.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "coding/traffic.hpp"
