+++ track_analyzing/track_analyzer/cmd_match.cpp
@@ -6,7 +6,7 @@
 
 #include "map/routing_helpers.hpp"
 
-#include "routing/num_mwm_id.hpp"
+#include "routing_common/num_mwm_id.hpp"
 
 #include "storage/country_info_getter.hpp"
 #include "storage/storage.hpp"
