+++ track_analyzing/track_matcher.hpp
@@ -3,8 +3,9 @@
 #include "track_analyzing/track.hpp"
 
 #include "routing/index_graph.hpp"
-#include "routing/num_mwm_id.hpp"
 #include "routing/segment.hpp"
+
+#include "routing_common/num_mwm_id.hpp"
 
 #include "indexer/index.hpp"
